Rearranges methods and wraps submit in a try-finally
