+++ lib/process.js
@@ -1,22 +1,124 @@
-const fs = require('fs');
+module.exports = function (self) {
+  return {
+    async submitForm (req) {
+      const input = JSON.parse(req.body.data);
+      const output = {};
+      const formErrors = [];
+      const formId = self.inferIdLocaleAndMode(req, input._id);
 
-module.exports = {
-  uploadFieldFiles: async function (req, name, files, {
-    warn,
-    insert
-  }) {
+      const form = await self.find(req, {
+        _id: self.apos.launder.id(formId)
+      }).toObject();
 
-    // Upload each file for the field, then join IDs.
-    try {
-      // Field names can't include
+      if (!form) {
+        throw self.apos.error('notfound');
+      }
+
+      if (form.enableRecaptcha) {
+        try {
+          // Process reCAPTCHA input if needed.
+          await self.checkRecaptcha(req, input, formErrors);
+        } catch (e) {
+          self.apos.util.error('reCAPTCHA submission error', e);
+          throw self.apos.error('invalid');
+        }
+      }
+
+      // Find any file field submissions and insert the files as attachments
+      for (const [ field, value ] of Object.entries(input)) {
+        if (value === 'files-pending') {
+          try {
+            input[field] = await self.uploadFieldFiles(req, field, req.files);
+          } catch (error) {
+            self.apos.util.error(error);
+            formErrors.push({
+              field: field,
+              error: 'invalid',
+              message: req.t('aposForm:fileUploadError')
+            });
+          }
+        }
+      }
+
+      // Recursively walk the area and its sub-areas so we find
+      // fields nested in two-column widgets and the like
+
+      // walk is not an async function so build an array of them to start
+      const areas = [];
+
+      self.apos.area.walk({
+        contents: form.contents
+      }, function(area) {
+        areas.push(area);
+      });
+
+      const fieldNames = [];
+      const conditionals = {};
+      const skipFields = [];
+
+      // Populate the conditionals object fully to clear disabled values
+      // before starting sanitization.
+      for (const area of areas) {
+        const widgets = area.items || [];
+        for (const widget of widgets) {
+          // Capture field names for the params check list.
+          fieldNames.push(widget.fieldName);
+
+          if (widget.type === '@apostrophecms/form-conditional') {
+            self.trackConditionals(conditionals, widget);
+          }
+        }
+      }
+
+      self.collectToSkip(input, conditionals, skipFields);
+
+      for (const area of areas) {
+        const widgets = area.items || [];
+        for (const widget of widgets) {
+          const manager = self.apos.area.getWidgetManager(widget.type);
+          if (
+            manager && manager.sanitizeFormField &&
+          !skipFields.includes(widget.fieldName)
+          ) {
+            try {
+              manager.checkRequired(req, widget, input);
+              await manager.sanitizeFormField(widget, input, output);
+            } catch (err) {
+              if (err.data && err.data.fieldError) {
+                formErrors.push(err.data.fieldError);
+              } else {
+                throw err;
+              }
+            }
+          }
+        }
+      }
+
+      if (formErrors.length > 0) {
+        throw self.apos.error('invalid', {
+          formErrors
+        });
+      }
+
+      if (form.enableQueryParams && form.queryParamList.length > 0) {
+        self.processQueryParams(form, input, output, fieldNames);
+      }
+
+      await self.emit('submission', req, form, output);
+
+      return {};
+    },
+    async uploadFieldFiles (req, name, files) {
+
+      // Upload each file for the field, then join IDs.
       const ids = [];
 
       for (const entry in files) {
-        if (!matchesName(entry, name)) {
+        if (!self.matchesName(entry, name)) {
           continue;
         }
 
-        const attachment = await insert(req, files[entry], {
+        const attachment = await self.apos.attachment.insert(req, files[entry], {
           permissions: false
         });
 
@@ -24,24 +126,51 @@ module.exports = {
       }
 
       return ids;
-    } finally {
-      for (const file of (Object.values(req.files || {}))) {
-        if (!matchesName(file.fieldName, name)) {
-          continue;
-        }
 
-        try {
-          fs.unlinkSync(file.path);
-        } catch (e) {
-          warn(req.t('aposForm:fileMissingEarly', {
-            path: file
-          }));
+    },
+    matchesName(str, name) {
+      return str.startsWith(name) && str.match(/.+-\d+$/);
+    },
+    trackConditionals(conditionals = {}, widget) {
+      const conditionName = widget.conditionName;
+      const conditionValue = widget.conditionValue;
+
+      if (!widget || !widget.contents || !widget.contents.items) {
+        return;
+      }
+
+      conditionals[conditionName] = conditionals[conditionName] || {};
+
+      conditionals[conditionName][conditionValue] = conditionals[conditionName][conditionValue] || [];
+
+      widget.contents.items.forEach(item => {
+        conditionals[conditionName][conditionValue].push(item.fieldName);
+      });
+
+      // If there aren't any fields in the conditional group, don't bother
+      // tracking it.
+      if (conditionals[conditionName][conditionValue].length === 0) {
+        delete conditionals[conditionName][conditionValue];
+      }
+    },
+    collectToSkip(input, conditionals, skipFields) {
+      // Check each field that controls a conditional group.
+      for (const name in conditionals) {
+        // For each value that a conditional group is looking for, check if the
+        // value matches in the output and, if not, remove the output properties
+        // for the conditional fields.
+        for (let value in conditionals[name]) {
+          // Booleans are tracked as true/false, but their field values are 'on'. TEMP?
+          if (input[name] === true && value === 'on') {
+            value = true;
+          }
+          if (input[name] !== value) {
+            conditionals[name][value].forEach(field => {
+              skipFields.push(field);
+            });
+          }
         }
       }
     }
-  }
+  };
 };
-
-function matchesName(str, name) {
-  return str.startsWith(name) && str.match(/.+-\d+$/);
-}
